prelabel/requirements.txt
certifi==2022.9.24
pika==1.3.1
pip==22.2.2
setuptools==63.4.1
wheel==0.37.1
opencv-python==4.6.0
